refactor: move ENS resolution into src/utils/ens.ts
Pure move, no behavior change: getEnsTextRecord, getEnsOwner and their
private helpers (getDomainType, getDNSOwner) and constants relocate from
utils.ts verbatim. utils.ts re-exports both, so the public API surface
is unchanged. Prepares a contained home for the upcoming ENS v2 /
Universal Resolver work (#1217 follow-up), so its functional diff stays
within one module.
